wip: Formular-widget_extraction_and_decomposition#467
